rename services files, thunkify app
