+++ src/js/components/chat-input.jsx
@@ -2,7 +2,7 @@ import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 import isMobile from 'ismobilejs';
 
-import { sendMessage, resetUnreadCount } from '../services/conversation-service';
+import { sendMessage, resetUnreadCount } from '../services/conversation';
 
 import { ImageUpload } from './image-upload';
