+++ src/js/components/header.jsx
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 
-import { toggleWidget, showSettings, hideSettings, hideChannelPage } from '../services/app-service';
+import { toggleWidget, showSettings, hideSettings, hideChannelPage } from '../services/app';
 import { hasChannels } from '../utils/app';
 import { CHANNEL_DETAILS } from '../constants/channels';
 import { WIDGET_STATE } from '../constants/app';
@@ -16,19 +16,28 @@ export class HeaderComponent extends Component {
     };
 
     showSettings(e) {
+        const {dispatch} = this.props;
         e.stopPropagation();
-        showSettings();
+        dispatch(showSettings());
     }
 
     hideSettings = (e) => {
         e.stopPropagation();
-        const {visibleChannelType} = this.props.appState;
+        const {dispatch, appState: {visibleChannelType}} = this.props;
         if (visibleChannelType) {
-            hideChannelPage();
+            dispatch(hideChannelPage());
         } else {
-            hideSettings();
+            dispatch(hideSettings());
         }
     }
+
+    onClick = (e) => {
+        e.preventDefault();
+        const {dispatch, appState:{embedded}} = this.props;
+        if(!embedded) {
+            dispatch(toggleWidget());
+        }
+    };
 
     render() {
         const {appState: {emailCaptureEnabled, settingsVisible, widgetState, embedded, visibleChannelType}, unreadCount, settings, text} = this.props;
@@ -88,7 +97,7 @@ export class HeaderComponent extends Component {
         }
 
         return <div id={ settingsMode ? 'sk-settings-header' : 'sk-header' }
-                    onClick={ !embedded && toggleWidget }
+                    onClick={ this.onClick }
                     className='sk-header-wrapper'
                     style={ style }>
                    { settingsButton }
