--- src/js/services/auth-service.js
@@ -1,5 +1,0 @@
-import { core } from './core';
-
-export function login(props) {
-    return core().appUsers.init(props);
-}
