+++ src/js/services/auth.js
@@ -1,0 +1,7 @@
+import { core } from './core';
+
+export function login(props) {
+    return (dispatch, getState) => {
+        return core(getState()).appUsers.init(props);
+    };
+}
