+++ src/js/services/core.js
@@ -2,12 +2,10 @@ import { Smooch } from 'smooch-core/lib/smooch';
 import urljoin from 'urljoin';
 
 import { VERSION } from '../constants/version';
-import { store } from '../stores/app-store';
 
-export function core() {
-    const auth = store.getState().auth;
+export function core({auth, appState}) {
     return new Smooch(auth, {
-        serviceUrl: urljoin(store.getState().appState.serverURL, 'v1'),
+        serviceUrl: urljoin(appState.serverURL, 'v1'),
         headers: {
             'x-smooch-sdk': `web/${VERSION}`
         }
