+++ test/specs/components/email-settings.spec.js
@@ -3,7 +3,7 @@ import TestUtils from 'react-addons-test-utils';
 import deepAssign from 'deep-assign';
 
 import { EmailSettings, EmailSettingsComponent } from '../../../src/js/components/email-settings';
-import * as userService from '../../../src/js/services/user-service';
+import * as userService from '../../../src/js/services/user';
 
 import { mockAppStore } from '../../utils/redux';
 import { wrapComponentWithStore } from '../../utils/react';
