+++ test/specs/components/header.spec.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import TestUtils from 'react-addons-test-utils';
 import deepAssign from 'deep-assign';
 
-import * as appService from '../../../src/js/services/app-service';
+import * as appService from '../../../src/js/services/app';
 import * as appUtils from '../../../src/js/utils/app';
 import { WIDGET_STATE } from '../../../src/js/constants/app';
 import { Header, HeaderComponent } from '../../../src/js/components/header';
