+++ test/specs/services/user-service.spec.js
@@ -4,8 +4,8 @@ import { createMock } from '../../mocks/core';
 import { mockAppStore } from '../../utils/redux';
 
 import * as coreService from '../../../src/js/services/core';
-import * as userService from '../../../src/js/services/user-service';
-import * as conversationService from '../../../src/js/services/conversation-service';
+import * as userService from '../../../src/js/services/user';
+import * as conversationService from '../../../src/js/services/conversation';
 
 describe('User service', () => {
     var sandbox;
